code/requirements.txt
pandas==2.2.3
pydantic==2.11.4
numpy==2.2.5
scikit-learn==1.6.1
sentence-transformers==4.1.0
llama-cpp-python==0.3.8
pytest==8.3.5
